+++ platform/src/pages/analytics/tabs/Explore.jsx
@@ -35,7 +35,8 @@ const Explore = () => {
           <CustomDropdown
             trigger={<TabButtons btnText={chartData.timeFrame} dropdown />}
             id='days'
-            dropStyle={{ top: '36px', zIndex: 9999, left: '0px' }}>
+            dropStyle={{ top: '36px', zIndex: 9999, left: '0px' }}
+          >
             {timeOptions.map((option) => (
               <span
                 key={option}
@@ -44,7 +45,8 @@ const Explore = () => {
                 }}
                 className={`cursor-pointer px-4 py-2 text-sm text-gray-700 hover:bg-gray-100 flex justify-between items-center ${
                   chartData.timeFrame === option ? 'bg-gray-100' : ''
-                }`}>
+                }`}
+              >
                 <span className='flex items-center space-x-2'>
                   <span>{option.charAt(0).toUpperCase() + option.slice(1)}</span>
                 </span>
@@ -57,7 +59,8 @@ const Explore = () => {
           <CustomDropdown
             trigger={<TabButtons Icon={BarChart} btnText='Chart' dropdown />}
             id='charts'
-            dropStyle={{ top: '36px', zIndex: 9999, right: '0px' }}>
+            dropStyle={{ top: '36px', zIndex: 9999, right: '0px' }}
+          >
             {chartOptions.map((option) => (
               <span
                 key={option.id}
@@ -66,7 +69,8 @@ const Explore = () => {
                 }}
                 className={`cursor-pointer px-4 py-2 text-sm text-gray-700 hover:bg-gray-100 flex justify-between items-center ${
                   chartData.chartType === option.id ? 'bg-gray-100' : ''
-                }`}>
+                }`}
+              >
                 <span className='flex items-center space-x-2'>
                   {option.icon}
                   <span>{option.name}</span>
@@ -84,6 +88,7 @@ const Explore = () => {
           chartType={chartData.chartType}
           chartTitle='Air quality over time'
           height={450}
+          id='explore-chart-container'
         />
       </div>
     </div>
